+++ cvlr-solana/src/layout/mod.rs
@@ -1,16 +1,18 @@
 mod common;
+#[cfg(feature = "rt")]
+pub(crate) mod instruction_accounts;
 #[cfg(not(feature = "rt"))]
 mod prover;
 #[cfg(feature = "rt")]
 mod rt;
 
+#[cfg(feature = "rt")]
+pub use instruction_accounts::InstructionAccounts;
 #[cfg(not(feature = "rt"))]
 pub use prover::{
     cvlr_deserialize_nondet_accounts, cvlr_deserialize_nondet_accounts as cvlr_nondet_acc_infos,
     cvlr_new_account_info, fun_acc_infos_with_mem_layout,
 };
-#[cfg(feature = "rt")]
-pub use rt::instruction_accounts::InstructionAccounts;
 #[cfg(feature = "rt")]
 pub use rt::{
     cvlr_deserialize_nondet_accounts, cvlr_deserialize_nondet_accounts as cvlr_nondet_acc_infos,
